requirements.txt
pyserial
pynput
numpy
psutils